requirements.txt
numpy>=1.26.0
pandas>=2.2.0
matplotlib>=3.8.0
seaborn>=0.13.0
scikit-learn>=1.4.0
cvxpy>=1.4.0
clarabel>=0.7.0
osqp>=0.6.3
torch>=2.2.0
pydantic>=2.6.0
pyyaml>=6.0
scipy>=1.12.0
